chore(server-route-and-controller): Separate routes from controllers This chore separates routes from controllers, Separates JOI validation framework from routes and refactors the server code to suit the seperated files. #Delivered
