+++ server.js
@@ -2,7 +2,6 @@
  * Created by obulaworld on 7/15/18.
  */
 import express from 'express';
-import path from 'path';
 import bodyParser from 'body-parser';
 // route to go here
 import router from './backend-routes/routes';
@@ -15,7 +14,8 @@ app.use(bodyParser.urlencoded({ extended: false }));
 
 // app.use(express.static(path.join(__dirname, 'UI')));
 
-app.use('/', router);
+// app.use('/api/v1', router);
+router(app);
 
 const server = app.listen(process.env.PORT || 3000, () => {
   console.log('Success!! app listening on port 3000!');
